+++ packages/jaspr_elementary_test/test/smoke_test.dart
@@ -2,7 +2,7 @@ import 'package:jaspr_elementary_test/jaspr_elementary_test.dart';
 import 'package:test/test.dart';
 
 void main() {
-  test('package loads', () {
-    jasprElementaryTestEnsureLinked();
+  test('library exports load', () {
+    expect(testComponentModel, isA<Function>());
   });
 }